implemented fucntionality to decompress module
