+++ src/commands/os/os.js
@@ -1,6 +1,8 @@
 import { EOL, cpus, homedir, userInfo, arch } from 'node:os';
 import { cwd } from 'node:process';
 
+import { ERRORS } from '../../utils/constants.js';
+
 export const os = (params) => {
   try {
     const flag = params.split(' ')[0];
@@ -11,7 +13,7 @@ export const os = (params) => {
       case '--cpus':
         const allCpus = cpus();
         console.log(`Overall amount of CPUS is ${allCpus.length}`);
-        allCpus.forEach((cpu, index) => console.log(`${index + 1} cpu: model:${cpu.model}, clock rate:${cpu.speed / 1000} GHz`));
+        allCpus.forEach((cpu, index) => console.log(`${index + 1}: model:${cpu.model}, clock rate:${cpu.speed / 1000} GHz`));
         break;
       case '--homedir':
         console.log(homedir());
@@ -24,11 +26,11 @@ export const os = (params) => {
         console.log(arch());
         break;
       default:
-        console.log('Enter the correct command');
+        console.error(ERRORS.INVALID_INPUT);
     };
 
     return console.log(`\nYou are currently in ${cwd()}`);
   } catch (err) {
-    console.error(err);
+    console.error(ERRORS.OPERATION_FAILED);
   }
 }
