+++ src/utils/commandProcessing.js
@@ -8,11 +8,12 @@ import { mv } from "../commands/fs/mv.js";
 import { rm } from "../commands/fs/rm.js";
 import { os } from "../commands/os/os.js";
 import { hash } from "../commands/hash/hash.js";
-import { compress } from "../commands/zip/zip.js";
+import { compress } from "../commands/zip/compress.js";
+import { decompress } from "../commands/zip/decompress.js";
 import { ERRORS } from "./constants.js";
 
 export const commandProcessing = async (command, params) => {
-  const commands = { cd, ls, rn, add, cat, cp, mv, rm, os, hash, compress };
+  const commands = { cd, ls, rn, add, cat, cp, mv, rm, os, hash, compress, decompress };
 
   if (commands[command]) {
     await commands[command](params);
